third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/style-attribute-expected.txt

CONSOLE ERROR: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-style-attr-value.pl?q=%22onload=%22alert(42)' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
